Consolidate mantis-source-jobs into this repo.
From https://github.com/Netflix/mantis-source-jobs.
