+++ mantis-source-jobs/mantis-source-job-kafka/build.gradle
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2019 Netflix, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+apply plugin: 'mantis'
+
+ext {
+    mqlVersion = '3.2.2'
+}
+
+dependencies {
+    implementation project(":mantis-runtime")
+    implementation project(":mantis-connectors:mantis-connector-kafka")
+
+    implementation "io.mantisrx:mql-jvm:$mqlVersion"
+
+    testCompile 'junit:junit'
+    testCompile "org.mockito:mockito-all:1.9.5"
+    // update kafka-unit version for kafka 2.2.+ once this PR is released https://github.com/chbatey/kafka-unit/pull/69
+    testCompile 'info.batey.kafka:kafka-unit:1.0'
+}
