+++ mantis-source-jobs/mantis-source-job-kafka/src/main/resources/META-INF/services/io.mantisrx.runtime.MantisJobProvider
