+++ mantis-source-jobs/mantis-source-job-kafka/src/test/java/io/mantisrx/sourcejob/kafka/QueryableKafkaSourceJobTest.java
