+++ mantis-source-jobs/mantis-source-job-publish/build.gradle
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2019 Netflix, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+apply plugin: 'mantis'
+
+ext {
+    junitVersion = '5.3.+'
+    mockitoVersion = '2.18.+'
+    slf4jVersion = '1.7.+'
+}
+
+dependencies {
+    implementation project(":mantis-runtime")
+    implementation project(":mantis-publish:mantis-publish-netty")
+    implementation project(":mantis-connectors:mantis-connector-publish")
+
+    implementation "org.slf4j:slf4j-api:$slf4jVersion"
+
+    testImplementation "org.junit.jupiter:junit-jupiter-api:$junitVersion"
+    testRuntimeOnly "org.junit.jupiter:junit-jupiter-engine:$junitVersion"
+    testImplementation "org.mockito:mockito-core:$mockitoVersion"
+}
+
+test {
+    useJUnitPlatform()
+}
