+++ mantis-source-jobs/mantis-source-job-publish/src/main/java/io/mantisrx/sourcejobs/publish/core/Utils.java
