+++ mantis-source-jobs/mantis-source-job-publish/src/main/resources/META-INF/services/io.mantisrx.runtime.MantisJobProvider
