--- temp-source-jobs/.github/CODEOWNERS
@@ -1,1 +1,0 @@
-* @codyrioux @neerajrj @nickmahilani @jeffchao @piygoyal @skalidindi
