--- temp-source-jobs/.github/ISSUE_TEMPLATE.md
@@ -1,15 +1,0 @@
-### Context
-
-Explain symptoms and other details for the issue.
-
-### Steps to reproduce
-
-Explain the steps to reliably reproduce the issue.
-
-### Expected behavior
-
-Explain what you think should happen.
-
-### Actual Behavior
-
-Explain what actually happens instead.
