--- temp-source-jobs/.github/PULL_REQUEST_TEMPLATE.md
@@ -1,11 +1,0 @@
-### Context
-
-Explain context and other details for this pull request.
-
-### Checklist
-
-- [ ] `./gradlew build` compiles code correctly
-- [ ] Added new tests where applicable
-- [ ] `./gradlew test` passes all tests
-- [ ] Extended README or added javadocs where applicable
-- [ ] Added copyright headers for new files from `CONTRIBUTING.md`
