--- temp-source-jobs/.travis.yml
@@ -1,14 +1,0 @@
-language: java
-jdk: openjdk8
-install: "./installViaTravis.sh"
-script: "./buildViaTravis.sh"
-cache:
-  directories:
-  - "$HOME/.gradle/caches/"
-  - "$HOME/.gradle/wrapper/"
-env:
-  global:
-  - secure: W8pR/aKtzO1inmrYkpBUzfxi5mmucFkvwiAGJnhZvzTf7utWWtjRKtDbFBwCCh/3SdvvO5icckvWgSuKbEM2jbIjNLXus9+XCDibhbzgel5uVjXsthKCbFlRL7hj/qFH3sskWEWobVUlQ9SE7NsdNJe9IW+OCc/qVpabRN6n4yZrqFtfFZKe9nsEvFWKre/F5of/nU3Fr6z9q64/0TylhW0+HP+CLnre0gwkNHE+WqMZTXYSGnqw+Hi+AAHolUMjzyL1fKOkieWVXecj/mm/JMCKmYWdyxYCE8DAU5BOYHJLYnT7Asy4VgC31GQazAQG026soTuMEwm5aZXMFgo0K7FzLmsVg1zxLzxrfqhqFICugPUajw4ZSmHEtz1uwTq/F/xFrHfz/tGTSfdI7dxCSINnJsxiRDzxo2Eux/RycB5cv3zcDGY2KeUr8aiIb0qNNi1BIQXd93NRhuMlknvGvGGMNkaT1PIM/BFrUM/DZIhaGwMEAOQ7211XR9zPI90oJfNi7YNGbIdtULAzJATk3id+5nq+h7ZCLOymtdkARJWcZlsMLI2lxUyw1xtaL4iqCjdM+qihakdE0V+fAqdylGe/EULxb784Aq8M5hImvVyHK4biYLYDlhJ6HYkq+HtVCJq7vOPsFQ4v09XJmqGyUS1hNu7U7FuEw8FnqJSn5dc=
-  - secure: ZVE9fx0oI7qO5+9PL03Tl9orSeMOOIUiWkLBh8DVGMTiVvPV+J48VxIuXY6HouA4M02hNHP+P4L4pD3+8kVCgTK10ba8hCRX9Ol6T5V/tRjl4OXEHeTFzhwm4vL5kUKqjLXYvWsqBOnggDKkEzX6aWloOIYB+2KHaS2yLfLrsCZe8QgZD5UqGsnk1KWbUyYGFvRFcEKnJbCqwvkjI3u6ehagMTdYrCHkzFTrKlrufwebac2HIH6HCMFhH7eGd5l8M8w7uF3Y0ME9GIh+JhWCCGp0H7kTPmDXiUKSaPCIg+WjXJj3UX4V8Ug1U8Hm9EWp1buOGExPdOqizNrBW8gbDGZ7BfANJ6EYpwx3u9lPcRq9OVnZjm3is10itSeGdzCZKTivkQ0Q6xfPeUt7wFT4EFct6py5XRdP2vz+4m2nuG/ANIXdu3k/WbectCSadQHoh8X+hLRr3kDpVNjrEffb5gFEyIO9SMYhzSqOLFUbWpCjuCkxOW9r5NqqN9nkhSRGLQNELxu3zst+7aFN5+gw/OyUuJf86r0+RJMth2lasJ9Un1wH/Zfyo7gKfFRGYoi/eReSLecHPlY770HnWYteKcQCaTX1IEUk+B2XkH12foOTXV/O2oL7Nw1l+CZtp3++yAOy1wYqWVoRUIxsRgb39FJTdDuzUb4nLBDYpChR0nw=
-  - secure: vf52lQMANIXNEb/2TWzcK5llZZU9gt6G7W2lXYT0V2jTA1nojRDGUuwQGFz884ZK94aKEsOBviu9NO64r1Bl2aBT5ZHz5aV49mOKjkBw5+1vmFfb4HZc99GmB8HRcZbrw0CVEw7Z0hAtnZ3UJYTTneoPkKxGGx7U1SEVolkZa+Vo3A6mwLBwhF1M6MFWosze3tjQSdsn7vwJVCKDx4YP89ERJYkHLKOyvFBq92VNg0wrSUv6Wio8oyJc2LnDF57YV8GFgTq3vHiYotLrW1D0YgaConirwoQpVi3693k696T2eaXAQrEIMN5r4lFbSKJ4ym2SLkDR9H1CPd2/cMD6aEFuUWpH0JWlfLeWKDS3cVwvkTZ/oTJOo11BSuQG7guTjaK+TGuwQxBwuUbkzCaTqLOFet09TAwhGDVGAZLOmzSAOVOQdZi3AWPwFF/uM+Osc3v4wgThy40ruutbJ7XaAESVRJNdgWFb84gYNWcY44FuNh9qBwcDggwp9OG6wsC4Zfy54iB4ydKkP6NfJ1x9+I1SVjwdkDxwIe0coZt8l1+7R2DL5ubBXipVDzKI3UhB67CWojfRN7rILTKcKjEZIQMayMO6D7RcOrAnW8xrN5Mu27Obh9hB1gpk1lLIeqrPxLgrlX3x64T10hQuDcH3Wxbf1F4I9lK42OnfVTg9Fzs=
-  - secure: f1EWbXuTXsd/geqXTU9JPYlx1P8g9eFhe+H0Avi9eZX22fidCGIruDF/xe+6irNsCKyL/eBAyinCBVP2e1vz9Am4bLM3c9h5ciF0SS8JKC4WLGj+IecHp7AV5ogQmK9+A0FVAcZlsSoQkvKzzaX1Oh+Pjv7sDdm2QNEy4BxOK6qrgolTvCAdmfkRAje68nDgY+wKQAobLjKsVsm5lRLrEweSu+ajcZaxxupj8CNpaqYTKFae6PouldkJjmAL27apJ/Qkar+FfIcZwA72OZit7e9C/JTTwVlWf/JldK8k42c6tp2MnOj6Dm0bnkZrU8WD5UA7JduzWGRXZB2NMqiYeuhj8t2vjLjwDzEIPnCr34JYZJ0HlMu8JOWZHBCD3Wo+D9/rUkg83JeUrTj5MZObeNqtZtwDFPOXzejYEciad9BUcDEbFh3k75SUvveIHfporxoDBpC0xTK9H8oV0bHZ/OHgTMRSlJuIwhnu4DOtrd9KYqs7jIKEnfce9EpCaH+2yjoFxH9lAJVoramaFCCIWEr+haJT4jLs40Tkdg6AoFXZL9VH70auYR8FXz54qRUjE6pEq/9uN4FJMnI01rwpdghPxNLTGHWZFI2AO/DtphdbyiTd5UxkDIBqdfxhz3n9NAR76/TNi5nKQ/TBB6J8xWkGBqskSkNV7vU+5AlA4AI=
