--- temp-source-jobs/OSSMETADATA
@@ -1,1 +1,0 @@
-osslifecycle=active
