--- temp-source-jobs/codequality/netflix-intellij-code-style.xml
@@ -1,108 +1,0 @@
-<!--
-  ~ Copyright 2019 Netflix, Inc.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<code_scheme name="Netflix" version="173">
-    <option name="OTHER_INDENT_OPTIONS">
-        <value>
-            <option name="CONTINUATION_INDENT_SIZE" value="0"/>
-        </value>
-    </option>
-    <option name="CLASS_COUNT_TO_USE_IMPORT_ON_DEMAND" value="99"/>
-    <option name="NAMES_COUNT_TO_USE_IMPORT_ON_DEMAND" value="99"/>
-    <option name="PACKAGES_TO_USE_IMPORT_ON_DEMAND">
-        <value/>
-    </option>
-    <option name="IMPORT_LAYOUT_TABLE">
-        <value>
-            <package name="" withSubpackages="true" static="true"/>
-            <emptyLine/>
-            <package name="java" withSubpackages="true" static="false"/>
-            <emptyLine/>
-            <package name="javax" withSubpackages="true" static="false"/>
-            <emptyLine/>
-            <package name="" withSubpackages="true" static="false"/>
-            <emptyLine/>
-            <emptyLine/>
-        </value>
-    </option>
-    <option name="RIGHT_MARGIN" value="100"/>
-    <option name="JD_ADD_BLANK_AFTER_PARM_COMMENTS" value="true"/>
-    <option name="JD_ADD_BLANK_AFTER_RETURN" value="true"/>
-    <GroovyCodeStyleSettings>
-        <option name="IMPORT_LAYOUT_TABLE">
-            <value>
-                <package name="" withSubpackages="true" static="true"/>
-                <emptyLine/>
-                <package name="" withSubpackages="true" static="false"/>
-                <emptyLine/>
-                <package name="java" withSubpackages="true" static="false"/>
-                <emptyLine/>
-                <package name="javax" withSubpackages="true" static="false"/>
-                <emptyLine/>
-                <emptyLine/>
-            </value>
-        </option>
-    </GroovyCodeStyleSettings>
-    <JavaCodeStyleSettings>
-        <option name="CLASS_NAMES_IN_JAVADOC" value="3"/>
-        <option name="CLASS_COUNT_TO_USE_IMPORT_ON_DEMAND" value="99"/>
-        <option name="NAMES_COUNT_TO_USE_IMPORT_ON_DEMAND" value="99"/>
-        <option name="PACKAGES_TO_USE_IMPORT_ON_DEMAND">
-            <value/>
-        </option>
-        <option name="IMPORT_LAYOUT_TABLE">
-            <value>
-                <package name="" withSubpackages="true" static="true"/>
-                <emptyLine/>
-                <package name="java" withSubpackages="true" static="false"/>
-                <emptyLine/>
-                <package name="javax" withSubpackages="true" static="false"/>
-                <emptyLine/>
-                <package name="" withSubpackages="true" static="false"/>
-                <emptyLine/>
-                <emptyLine/>
-            </value>
-        </option>
-        <option name="JD_ADD_BLANK_AFTER_PARM_COMMENTS" value="true"/>
-        <option name="JD_ADD_BLANK_AFTER_RETURN" value="true"/>
-    </JavaCodeStyleSettings>
-    <XML>
-        <option name="XML_LEGACY_SETTINGS_IMPORTED" value="true"/>
-    </XML>
-    <codeStyleSettings language="Groovy">
-        <indentOptions>
-            <option name="CONTINUATION_INDENT_SIZE" value="4"/>
-        </indentOptions>
-    </codeStyleSettings>
-    <codeStyleSettings language="HTML">
-        <indentOptions>
-            <option name="CONTINUATION_INDENT_SIZE" value="0"/>
-        </indentOptions>
-    </codeStyleSettings>
-    <codeStyleSettings language="JAVA">
-        <option name="RIGHT_MARGIN" value="120"/>
-        <option name="KEEP_FIRST_COLUMN_COMMENT" value="false"/>
-        <option name="BLANK_LINES_AFTER_IMPORTS" value="2"/>
-        <option name="BLANK_LINES_AFTER_CLASS_HEADER" value="1"/>
-        <option name="INDENT_CASE_FROM_SWITCH" value="false"/>
-        <option name="SPACE_BEFORE_ARRAY_INITIALIZER_LBRACE" value="true"/>
-        <option name="SPACE_BEFORE_ANNOTATION_ARRAY_INITIALIZER_LBRACE" value="true"/>
-        <option name="BINARY_OPERATION_SIGN_ON_NEXT_LINE" value="true"/>
-        <option name="KEEP_SIMPLE_BLOCKS_IN_ONE_LINE" value="true"/>
-        <option name="KEEP_SIMPLE_METHODS_IN_ONE_LINE" value="true"/>
-        <option name="KEEP_SIMPLE_CLASSES_IN_ONE_LINE" value="true"/>
-    </codeStyleSettings>
-</code_scheme>
