--- temp-source-jobs/codequality/pmd.xml
@@ -1,97 +1,0 @@
-<?xml version="1.0"?>
-<!--
-  ~ Copyright 2019 Netflix, Inc.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<ruleset xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         name="Custom ruleset"
-         xmlns="http://pmd.sf.net/ruleset/1.0.0"
-         xsi:schemaLocation="http://pmd.sf.net/ruleset/1.0.0 http://pmd.sf.net/ruleset_xml_schema.xsd"
-         xsi:noNamespaceSchemaLocation="http://pmd.sf.net/ruleset_xml_schema.xsd">
-    <description>
-        Exclude noisy rules.
-    </description>
-
-    <rule ref="rulesets/internal/all-java.xml">
-        <exclude name="AccessorMethodGeneration"/>
-        <exclude name="AddEmptyString"/>
-        <exclude name="AppendCharacterWithChar"/>
-        <exclude name="ArrayIsStoredDirectly"/>
-        <exclude name="AtLeastOneConstructor"/>
-        <exclude name="AvoidCatchingGenericException"/>
-        <exclude name="AvoidDuplicateLiterals"/>
-        <exclude name="AvoidFieldNameMatchingMethodName"/>
-        <exclude name="AvoidFinalLocalVariable"/>
-        <exclude name="AvoidInstantiatingObjectsInLoops"/>
-        <exclude name="AvoidLiteralsInIfCondition"/>
-        <exclude name="AvoidSynchronizedAtMethodLevel"/>
-        <exclude name="AvoidThrowingNullPointerException"/>
-        <exclude name="AvoidThrowingRawExceptionTypes"/>
-        <exclude name="AvoidUsingShortType"/>
-        <exclude name="AvoidUsingVolatile"/>
-        <exclude name="BeanMembersShouldSerialize"/>
-        <exclude name="CallSuperInConstructor"/>
-        <exclude name="ClassNamingConventions"/>
-        <exclude name="CommentDefaultAccessModifier"/>
-        <exclude name="CommentRequired"/>
-        <exclude name="CommentSize"/>
-        <exclude name="CompareObjectsWithEquals"/>
-        <exclude name="ConfusingTernary"/>
-        <exclude name="ControlStatementBraces"/>
-        <exclude name="CyclomaticComplexity"/>
-        <exclude name="DataflowAnomalyAnalysis"/>
-        <exclude name="DefaultPackage"/>
-        <exclude name="DoNotUseThreads"/>
-        <exclude name="DontImportJavaLang"/> <!-- Incorrectly flags java.lang.invoke.* -->
-        <exclude name="DuplicateImports"/>
-        <exclude name="EmptyMethodInAbstractClassShouldBeAbstract"/>
-        <exclude name="ExcessiveImports"/>
-        <exclude name="ExcessivePublicCount"/>
-        <exclude name="FieldDeclarationsShouldBeAtStartOfClass"/>
-        <exclude name="FieldNamingConventions"/>
-        <exclude name="GodClass"/>
-        <exclude name="GuardLogStatement"/>
-        <exclude name="ImmutableField"/>
-        <exclude name="LawOfDemeter"/>
-        <exclude name="LocalVariableCouldBeFinal"/>
-        <exclude name="LoggerIsNotStaticFinal"/>
-        <exclude name="LongVariable"/>
-        <exclude name="LoosePackageCoupling"/>
-        <exclude name="MethodArgumentCouldBeFinal"/>
-        <exclude name="MethodReturnsInternalArray"/>
-        <exclude name="NullAssignment"/>
-        <exclude name="OneDeclarationPerLine"/>
-        <exclude name="OnlyOneReturn"/>
-        <exclude name="PrematureDeclaration"/>
-        <exclude name="RedundantFieldInitializer"/>
-        <exclude name="SignatureDeclareThrowsException"/>
-        <exclude name="SimpleDateFormatNeedsLocale"/>
-        <exclude name="SimplifyConditional"/>
-        <exclude name="ShortClassName"/>
-        <exclude name="ShortMethodName"/>
-        <exclude name="ShortMethodName"/>
-        <exclude name="ShortVariable"/>
-        <exclude name="TooManyFields"/>
-        <exclude name="TooManyMethods"/>
-        <exclude name="UncommentedEmptyMethodBody"/>
-        <exclude name="UnusedPrivateMethod"/>
-        <exclude name="UnnecessaryConstructor"/>
-        <exclude name="UseConcurrentHashMap"/>
-        <exclude name="UseIndexOfChar"/>
-        <exclude name="UselessParentheses"/>
-        <exclude name="UseStringBufferForStringAppends"/>
-        <exclude name="UseVarargs"/>
-    </rule>
-</ruleset>
