--- temp-source-jobs/gradle/check.gradle
@@ -1,34 +1,0 @@
-/*
- * Copyright 2019 Netflix, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-subprojects {
-    checkstyle {
-        toolVersion = '8.14'
-        // TODO: Don't ignore failures.
-        ignoreFailures = true
-        configFile = rootProject.file('codequality/checkstyle.xml')
-        sourceSets = [sourceSets.main]
-    }
-
-    pmd {
-        toolVersion = '6.9.0'
-        // TODO: Don't ignore failures.
-        ignoreFailures = true
-        sourceSets = [sourceSets.main]
-        ruleSets = []
-        ruleSetFiles = rootProject.files("codequality/pmd.xml")
-    }
-}
